Splitting the classes to their own files.
