+++ starter_kits/TypeScript/hlt/Direction.ts
@@ -1,4 +1,4 @@
-import { Commands } from './commands';
+import { Commands } from './Commands';
 
 export class Direction {
     static North = new Direction(0, -1);
@@ -60,51 +60,5 @@ export class Direction {
             return Direction.Still;
         }
         throw new Error(`Non-cardinal direction cannot be inverted: ${this}`);
-    }
-}
-
-export class Position {
-    constructor(public x: number, public y: number) {
-        this.x = x;
-        this.y = y;
-    }
-
-    directionalOffset(direction: Direction) {
-        return this.add(new Position(direction.dx, direction.dy));
-    }
-
-    getSurroundingCardinals() {
-        return Direction.getAllCardinals()
-            .map(currentDirection => this.directionalOffset(currentDirection));
-    }
-
-    add(other: Position) {
-        return new Position(this.x + other.x, this.y + other.y);
-    }
-
-    sub(other: Position) {
-        return new Position(this.x - other.x, this.y - other.y);
-    }
-
-    addMut(other: Position) {
-        this.x += other.x;
-        this.y += other.y;
-    }
-
-    subMut(other: Position) {
-        this.x -= other.x;
-        this.y -= other.y;
-    }
-
-    abs() {
-        return new Position(Math.abs(this.x), Math.abs(this.y));
-    }
-
-    equals(other: Position) {
-        return this.x === other.x && this.y === other.y;
-    }
-
-    toString() {
-        return `${this.constructor.name}(${this.x}, ${this.y})`;
     }
 }
