+++ starter_kits/TypeScript/hlt/gameMap.ts
@@ -1,144 +1,10 @@
-import { Shipyard, Ship, Dropoff, Entity } from "./entity";
-import { Position, Direction } from "./positionals";
-import { Constants } from "./constants";
+import { Entity } from "./Entity";
+import { Ship } from "./Ship";
+import { Direction } from "./Direction";
+import { Position } from "./Position";
+import { Constants } from "./Constants";
+import { MapCell } from "./MapCell";
 
-
-/** Player object, containing all entities/metadata for the player. */
-export class Player {
-    private _ships = new Map<number, Ship>();
-    private _dropoffs = new Map<number, Dropoff>();
-
-    constructor(
-        public id: number,
-        public shipyard: Shipyard,
-        public haliteAmount = 0
-    ) {
-    }
-
-    /** Get a single ship by its ID. */
-    getShip(shipId: number) {
-        return this._ships.get(shipId);
-    }
-
-    /** Get a list of the player's ships. */
-    getShips() {
-        const result = [];
-        for (const ship of this._ships.values()) {
-            result.push(ship);
-        }
-        return result;
-    }
-
-    /** Get a single dropoff by its ID. */
-    getDropoff(dropoffId: number) {
-        return this._dropoffs.get(dropoffId);
-    }
-
-    /** Get a list of the player's dropoffs. */
-    getDropoffs() {
-        const result = [];
-        for (const dropoff of this._dropoffs.values()) {
-            result.push(dropoff);
-        }
-        return result;
-    }
-
-    /** Check whether a ship with a given ID exists. */
-    hasShip(shipId: number) {
-        return this._ships.has(shipId);
-    }
-
-    /**
-     * Create a player object using input from the game engine.
-     */
-    static async _generate(getLine: () => Promise<string>) {
-        const line = await getLine();
-        const [playerId, shipyardX, shipyardY] = line
-            .split(/\s+/)
-            .map(x => parseInt(x, 10));
-        return new Player(playerId,
-            new Shipyard(playerId, -1, new Position(shipyardX, shipyardY)));
-    }
-
-    /**
-     * Update the player object for the current turn using input from
-     * the game engine.
-     * @private
-     */
-    async _update(numShips: number, numDropoffs: number, halite: number, getLine: () => Promise<string>) {
-        this.haliteAmount = halite;
-        this._ships = new Map();
-        for (let i = 0; i < numShips; i++) {
-            const { shipId, ship } = await Ship._generate(this.id, getLine);
-            this._ships.set(shipId, ship);
-        }
-        this._dropoffs = new Map();
-        for (let i = 0; i < numDropoffs; i++) {
-            const { id, dropoff } = await Dropoff._generate(this.id, getLine);
-            this._dropoffs.set(id, dropoff);
-        }
-    }
-}
-
-/** A cell on the game map. */
-export class MapCell {
-    ship: Ship | null = null;
-    structure: any = null;
-
-    constructor(
-        public position: Position,
-        public haliteAmount: number
-    ) {
-    }
-
-    /**
-     * Whether this cell has no ships or structures.
-     */
-    get isEmpty() {
-        return !this.isOccupied && !this.hasStructure;
-    }
-
-    /**
-     * Whether this cell has any ships.
-     */
-    get isOccupied() {
-        return this.ship !== null;
-    }
-
-    /**
-     * Whether this cell has any structures.
-     */
-    get hasStructure() {
-        return this.structure !== null;
-    }
-
-    /**
-     * @returns The type of the structure in this cell, or null.
-     */
-    get structureType() {
-        if (this.structure !== null) {
-            return this.structure.constructor;
-        }
-        return null;
-    }
-
-    /**
-     * Mark this cell as unsafe (occupied) for navigation.
-     *
-     * Use in conjunction with {@link GameMap#getSafeMove}.
-     */
-    markUnsafe(ship: Ship) {
-        this.ship = ship;
-    }
-
-    equals(other: MapCell) {
-        return this.position.equals(other.position);
-    }
-
-    toString() {
-        return `MapCell(${this.position}, halite=${this.haliteAmount})`;
-    }
-}
 
 /**
  * The game map.
